requirements.txt
environs==9.5.0
grpcio==1.58.0
matplotlib==3.8.2
milvus==2.2.16
minio==7.2.0
numpy==1.26.3
pandas==2.1.4
Pillow==10.1.0
Pillow==10.2.0
protobuf==4.25.1
Requests==2.31.0
setuptools==65.5.0
torch==2.1.1
torchinfo==1.8.0
torchvision==0.16.1
tqdm==4.66.1
ujson==5.8.0
util==0.0.1
